DiscretisedFluidQueues

Numerical approximation schemes for stochastic fluid queues.

The evolution of a stochastic fluid queue with generator $[T_{ij}]_{i,j\in S}$ and associated diagonal matrix of rates $C = diag(c_i,i\in S)$ can be described by the PDE (when it exists)

\[\cfrac{\partial}{\partial t} \mathbf f(x,i,t) = \mathbf f(x,i,t) T - \cfrac{\partial}{\partial x} \mathbf f(x,i,t) C\]

where $\mathbf f(x,i,t) dx = (f(x,i,t))_{i\in S} = (P(X(t)\in dx, \varphi(t)=i))_{i\in S}$ the time-dependent joint density/mass function.

This package implements finite element and finite volume numerical solvers to approximate the right-hand side of this PDE;

Discontinuous Galerkin: projects the right-hand side of the PDE on to a basis of polynomials,

Finite volume:

QBD-RAP approximation: use matrix-exponential distributions to model the solution locally on each cell.

Usage

pkg> add https://github.com/angus-lewis/DiscretisedFluidQueues

julia> using DiscretisedFluidQueues

Create a model with (for example)

T = [-2.5 2 0.5; 1 -2 1; 1 2 -3] # generator of the phase
C = [0.0; 2.0; -3.0]    # rates dX/dt

S = PhaseSet(C) 

model = FluidQueue(T,S) 

Create a discretisation mesh (a grid + method with which to approximate the solution) with any of (e.g.)

nbases = 3
mesh = DGMesh(nodes,nbases)

fv_order = 3
fvmesh = FVMesh(nodes,fv_order)

order = 3
frapmesh = FRAPMesh(nodes,order)

Combine the model and the discretisation scheme (mesh) to form a discretised fluid queue

dq = DiscretisedFluidQueue(model,mesh)

Construct an approximation to the generator with

B = build_full_generator(dq)

B is essentially a matrix which we can think of as describing the ODE \begin{equation} \frac{\partial}{\partial t} \mathbf a(t) = \mathbf a(t) B \end{equation} where $ \mathbf a(t)$ is a row vector of coefficients and $ \mathbf a(t) \mathbf u(x,i) \approx \mathbf f(x,i,t)$ approximates the solution where $ \mathbf u(x,i)$ is a column vector of functions defined by the discretisation scheme.

Construct an initial distribution with (e.g.)

f(x,i) = (i-1)/12.0./sum(1:3) # the initial distribution
d = SFMDistribution(f,dq)

Integrate over time with

t = 3.2
dt = integrate_time(d,B,t)

Reconstruct approximate solution with

u = cdf(d)

Evaluate solution as a function

u(0.1,1) # = P(X(3.2)<0.1, phi(3.2)=1)

More examples and documentation to come.

This is my PhD topic. I'll provide a link to a write up when I've done one...

DiscretisedFluidQueues.N₊ — Method

N₊(S::PhaseSet) = begin

Return the number of right/upper point masses of a PhaseSet, FluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.N₋ — Method

N₋(S::PhaseSet)

Return the number of left/lower point masses of a PhaseSet, FluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.augment_model — Method

augment_model(model::FluidQueue)

Given a FluidQueue, return a FluidQueue with twice as many phases with rate 0, one set associated with m=1 phases and one associated with m=-1 phases.

DiscretisedFluidQueues.build_full_generator — Method

build_full_generator(dq::DiscretisedFluidQueue; v::Bool = false)

Returns a SparseMatrixCSC generator of a DiscretisedFluidQueue.

DiscretisedFluidQueues.build_full_generator — Method

build_full_generator(lzB::LazyGenerator)

Returns a SparseMatrixCSC representation of a LazyGenerator.

DiscretisedFluidQueues.build_lazy_generator — Method

build_lazy_generator(dq::DiscretisedFluidQueue; v::Bool = false)

Build a lazy representation of the generator of a discretised fluid queue.

DiscretisedFluidQueues.build_me — Method

build_me(params; mean::Real = 1)

Return a ConcentratedMatrixExponential as defined by the dictionary params. i.e. see CMEParams[3] for an order 3 CME.

DiscretisedFluidQueues.ccdf — Method

ccdf(a::Array{Float64, 2}, me::AbstractMatrixExponential, x::Real)

Return a the complimentary cumulative distribution function (1-cdf) of a matrix exponential evaluated at x.

DiscretisedFluidQueues.ccdf — Method

ccdf(a::Array{Float64, 2}, me::AbstractMatrixExponential)

Return a the complimentary cumulative distribution function (1-cdf) of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)

DiscretisedFluidQueues.cdf — Method

cdf(a::Array{Float64, 2}, me::AbstractMatrixExponential, x::Real)

Return a the cumulative distribution function of a matrix exponential evaluated at x.

DiscretisedFluidQueues.cdf — Method

cdf(a::Array{Float64, 2}, me::AbstractMatrixExponential) = begin

Return a the cumulative distribution function of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)

DiscretisedFluidQueues.cdf — Method

cdf(d::SFMDistribution, x, i)

Evaluate cdf(d::SFMDistribution) at (x,i).

DiscretisedFluidQueues.cdf — Method

cdf(d::SFMDistribution)

Return a function of two variables (x::Float64,i::Int) which is the cumulative distribution distribution function defined by d. Approximates the distribution function of a fluid queue.

DiscretisedFluidQueues.cdf — Method

cdf(s::Simulation)

Return a empirical cdf of the fluid queue given a simulation.

Output is a function of two variables (x,i) and gives the empirical distribution function $P(X(τ)≤x,φ(τ)=i)$

DiscretisedFluidQueues.cell_nodes — Method

cell_nodes(mesh::DGMesh)

The positions of the GLJ nodes within each cell of a mesh.

DiscretisedFluidQueues.cell_nodes — Method

cell_nodes(mesh::FRAPMesh)

The cell centre

DiscretisedFluidQueues.cell_nodes — Method

cell_nodes(mesh::FVMesh)

The cell centres

DiscretisedFluidQueues.first_exit_x — Method

Constructs the StoppingTime which is the first exit of the process $X(t)$ from the interval $[u,v]$.

first_exit_x( u::Real, v::Real)

Arguments

u: a lower boundary

v: an upper boundary

Output

first_exit_xFun: a function with one methods

`firstexitxFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.

DiscretisedFluidQueues.fixed_time — Method

Constructs the StoppingTime $1(t>T)$

fixed_time( T::Real)

Arguments

T: a time at which to stop the process

Output

fixed_timeFun: a function with one methods

`fixed_timeFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.

DiscretisedFluidQueues.integrate_time — Function

Given x0 apprximate x0 exp(Dy).

integrate_time(x0::AbstractArray{<:Real, 2}, D::Generator, y::Real, h::TimeIntegrationScheme = Euler(y ./ 1000))

Arguments

x0: An initial row vector

D: A square matrix

y: time to integrate up to

h: TimeIntegrationScheme.

Output

DiscretisedFluidQueues.interior_point_mass — Method

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{DGMesh})

Constructs a polynomial approximation to the point mass at (x,i)

DiscretisedFluidQueues.interior_point_mass — Method

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{FRAPMesh})

Return the appropriate initial condition to approximate the initial distribution which is a point mass at (x,i) for the numerical discretisation scheme defined by the dq DiscretisedFluidQueue.

i.e. Compute the orbit position $a(x) = a exp(S(x-yₖ))$ if the membership of i is -1 and $a(x) = a exp(S(yₖ₊₁-x))$ if the membership of i is 1 where yₖ and yₖ₊₁ are the left and right cell edges and x∈[yₖ,yₖ₊₁], respectively, and a and S are defined by dq.mesh.me are parameters of a MatrixExponential.

DiscretisedFluidQueues.interior_point_mass — Method

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{FVMesh})

Construct an approximation to a point mass at (x,i). Basically, just puts mass 1 in the cell containing (x,i).

DiscretisedFluidQueues.interior_point_mass — Method

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{<:Mesh})

Construct an approximation to a distribution of fluid queue with a point mass at (x,i). Returns a SFMDistribution. The method of approximation is defined by the type of <:Mesh.

Arguments:

x::Float64: The position of the point mass

i::Int: The phase of the point mass

dq::DiscretisedFluidQueue:

DiscretisedFluidQueues.left_point_mass — Method

left_point_mass(i::Int, dq::DiscretisedFluidQueue)

Construct a SFMDistribution with a point mass at the left boundary of dq in phase i and 0 elsewhere.

DiscretisedFluidQueues.membership — Method

membership(S::PhaseSet, i::Int)

Return the membership of phases from a PhaseSet

DiscretisedFluidQueues.n_bases_per_cell — Method

n_bases_per_cell(mesh::DGMesh)

Number of bases in a cell

DiscretisedFluidQueues.n_bases_per_cell — Method

n_bases_per_cell(m) =

Return the number of basis functions on each cell of a given DiscretisedFluidQueue or Mesh.

DiscretisedFluidQueues.n_bases_per_cell — Method

n_bases_per_cell(mesh::FRAPMesh)

Number of bases in a cell

DiscretisedFluidQueues.n_bases_per_cell — Method

n_bases_per_cell(mesh::FVMesh)

Constant 1

DiscretisedFluidQueues.n_bases_per_phase — Method

n_bases_per_phase(m)

Total number of basis functions used to represent the fluid queue for each phase of a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.n_intervals — Method

n_intervals(m)

Total number of cells for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.n_jumps — Method

Constructs the StoppingTime $1(N(t)>n)$ where $N(t)$ is the number of jumps of $φ$ by time $t$.

n_jumps( N::Int)

Arguments

N: a desired number of jumps

Output

n_jumpsFun: a function with one methods

`n_jumpsFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.

DiscretisedFluidQueues.n_phases — Method

n_phases(S::PhaseSet)

Return the number of phases from a PhaseSet, FluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.pdf — Method

pdf(a::Array{Float64, 2}, me::AbstractMatrixExponential, x::Real)

Return a the probability density function of a matrix exponential evaluated at x.

DiscretisedFluidQueues.pdf — Method

pdf([a::Array{Float64, 2},] me::AbstractMatrixExponential)

Return a the probability density function of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)

DiscretisedFluidQueues.pdf — Method

pdf(d::SFMDistribution, x, i)

Evaluate pdf(d::SFMDistribution) at (x,i).

DiscretisedFluidQueues.pdf — Method

pdf(d::SFMDistribution)

Return a function of two variables (x::Float64,i::Int) which is the probability distribution function defined by d. Approximates the density function of a fluid queue.

DiscretisedFluidQueues.phases — Method

phases(S::PhaseSet)

Return the iterator 1:n_phases(S), for a PhaseSet, FluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.psi_fun_x — Method

Construct and evaluate $Ψ(s)$ for a triditional SFM.

Uses newtons method to solve the Ricatti equation $D⁺⁻(s) + Ψ(s)D⁻⁺(s)Ψ(s) + Ψ(s)D⁻⁻(s) + D⁺⁺(s)Ψ(s) = 0.$

psi_fun_x( model::Model; s = 0, MaxIters = 1000, err = 1e-8)

Arguments

model: a Model object

s::Real: a value to evaluate the LST at

MaxIters::Int: the maximum number of iterations of newtons method

err::Float64: an error tolerance for terminating newtons method. Terminates when max(Ψ_{n} - Ψ{n-1}) .< eps.

Output

Ψ(s)::Array{Float64,2} the matrix $Ψ$

DiscretisedFluidQueues.rates — Method

rates(S::PhaseSet, i::Int)

Return the rate of the phase from a PhaseSet, FluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.right_point_mass — Method

right_point_mass(i::Int, dq::DiscretisedFluidQueue)

Construct a SFMDistribution with a point mass at the left boundary of dq in phase i and 0 elsewhere.

DiscretisedFluidQueues.simulate — Function

Simulates a SFM defined by Model until the StoppingTime has occured, given the InitialCondition on (φ(0),X(0)).

simulate(
    model::Model,
    lwr::Float64, upr::Float64,
    StoppingTime::Function,
    InitCondition::NamedTuple{(:φ, :X)},
)

Arguments

model: A Model object

lwr and upr upper an lower regulated boundaries for the fluid level.

StoppingTime: A function which takes the value of the process at the current time and at the time of the last jump of the phase process, as well as the model object. i.e. StoppingTime(;model,SFM,SFM0,lwer,upr) where SFM and SFM0 are tuples with keys (:t::Float64, :φ::Int, :X::Float64, :n::Int) which are the value of the SFM at the current time, and time of the previous jump of the phase process, repsectively. The StoppingTime must return a NamedTuple{(:Ind, :SFM)} type where :Ind is a :Bool value stating whether the stopping time has occured or not and :SFM is a tuple in the same form as the input SFM but which contains the value of the SFM at the stopping time.

InitCondition: NamedTuple with keys (:φ, :X), InitCondition.φ is a vector of length M of initial states for the phase, and InitCondition.X is a vector of length M of initial states for the level. M is the number of simulations to be done.

Output

A Simulation object.

DiscretisedFluidQueues.stationary_distribution_x — Method

Construct the stationary distribution of the SFM

stationary_distribution_x( model::Model, Ψ::Array, ξ::Array)

Arguments

model: a Model object

Ψ: an array as output from psi_fun_x

ξ: an array as returned from xi_x

Output

pₓ::Array{Float64,2}: the point masses of the SFM

πₓ(x) a function with two methods

πₓ(x::Real): for scalar inputs, returns the stationary density evaluated at x in all phases.

πₓ(x::Array): for array inputs, returns an array with the same shape as is output by Coeff2Dist.

K::Array{Float64,2}: the matrix in the exponential of the density.

DiscretisedFluidQueues.total_n_bases — Method

total_n_bases(m)

Return the number of basis functions of a given DiscretisedFluidQueue or Mesh.

DiscretisedFluidQueues.xi_x — Method

Construct the vector $ξ$ containing the distribution of the phase at the time when $X(t)$ first hits 0.

xi_x( model::Model, Ψ::Array)

Arguments

model: a Model object

Ψ: an array as output from psi_fun_x

Output

the vector ξ

DiscretisedFluidQueues.Δ — Method

Δ(m,k::Int)

The width of cell k for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.Δ — Method

Δ(m)

The width of each cell for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.ConcentratedMatrixExponential — Type

Similar to MatrixExponential but has a neater implementation of orbit() due to nice structure.

DiscretisedFluidQueues.DGMesh — Type

DGMesh <: Mesh

A structure representing a discretisation scheme to be used for a DiscretisedFluidQueue.

Arguments:

nodes::Array{<:Real, 1}: The edges of the cells.

n_bases::Int: The number of basis functions used to represent the solution on each cell

DiscretisedFluidQueues.DiscretisedFluidQueue — Type

DiscretisedFluidQueue{T <: Mesh}

A container to hold a FluidQueue and a discretisation scheme (a mesh). Parametric type depending on the type of discretisation scheme (<:Mesh)

Arguments:

model::Model: a FluidQueue

mesh::T: either a DGMesh, FVMesh or FRAPMesh specifying the discretisation scheme for the continuous level of the fluid queue, X(t).

DiscretisedFluidQueues.Euler — Type

Euler <: TimeIntegrationScheme

Defines an Euler integration scheme to be used in integrate_time.

Arguments

step_size::Float64: the step size of the integration scheme.

DiscretisedFluidQueues.FRAPMesh — Type

FRAPMesh <: Mesh

A QBD-RAP (FRAM) discretisation scheme for a DiscretisedFluidQueue.

Arguments

nodes: the cell edges

me: the MatrixExponential used to approximate model the fluid queue on each cell.

DiscretisedFluidQueues.FVMesh — Type

FVMesh <: Mesh

A finite volume discretisation scheme for a DiscretisedFluidQueue.

Arguments

nodes: the cell edges (the nodes are at the center of the cell)

order: the order of the polynomial interpolation used to approximate the flux at the

cell edges.

DiscretisedFluidQueues.FluidQueue — Type

FluidQueue <: Model

Constructor for a fluid queue model.

Arguments:

T::Array{<:Real, 2}: Generator of the phase process

S::PhaseSet: An array of phases describing the evolution of the fluid level in each phase.

DiscretisedFluidQueues.FullGenerator — Type

FullGenerator <: Generator

An Matrix or SparseMatrixCSC representation of the generator of a generator of a DiscretisedFluidQueue.

Higher memory requirements than LazyGenerator (as blocks are duplicated) but much faster matrix arithmetic.

DiscretisedFluidQueues.Generator — Type

Generator

Abstract type representing a discretised infinitesimal generator of a FLuidQueue.

DiscretisedFluidQueues.LazyGenerator — Type

LazyGenerator <: Generator

A lazy representation of a block matrix with is a generator of a DiscretisedFluidQueue.

Lower memory requirements than FullGenerator but aritmetic operations and indexing may be slower.

Arguments:

dq::DiscretisedFluidQueue:

blocks::Tuple{Array{Float64, 2}, Array{Float64, 2}, Array{Float64, 2}, Array{Float64, 2}}: Block matrices describing the flow of mass within and between cells. blocks[1] is the lower diagonal block describing the flow of mass from cell k+1 to cell (k for phases with negative rate only). blocks[2] (blocks[3]) is the diagonal block describing the flow of mass within a cell for a phase with positive (negative) rate. blocks[4] is the upper diagonal block describing the flow of mass from cell k to k+1 (for phases with positive rate only).

boundary_flux::BoundaryFluxTupleType: A named tuple structure such that - boundary_flux.lower.in: describes flow of density into lower boundary - boundary_flux.lower.out: describes flow of density out of lower boundary - boundary_flux.upper.in: describes flow of density into upper boundary - boundary_flux.upper.out: describes flow of density out of upper boundary

D::Union{Array{Float64, 2}, LinearAlgebra.Diagonal{Bool, Array{Bool, 1}}}: An array describing how the flow of density changes when the phase process jumps between phases with different memberships. This is the identity for FV and DG schemes.

DiscretisedFluidQueues.MatrixExponential — Type

MatrixExponential constructor method

MatrixExponential(
    a::Array{Float64,2},
    S::Array{Float64,2},
    s::Array{Float64,1},
    D::Array{Float64}=[0],
)

Inputs:

a a 1 by p Array of reals

S a p by p Array of reals

s a p by 1 Array of reals

D an optional argument, if empty then me.D is the identity, else is a p by p matrix

Throws an error if the dimensions are inconsistent.

DiscretisedFluidQueues.Mesh — Type

Mesh

Abstract type representing a discretisation mesh for a numerical scheme.

DiscretisedFluidQueues.Phase — Type

Phase

Represents a phase of the fluid queue.

Arguments:

c::Float64: the rate of change of the fluid X(t) associated with the phase

m::Int64: either -1 or 1, membership of the phase with either the set of positive or negative phases. Default is positive, i.e. 1. Only relevant for FRAPApproximation.

lpm::Bool: whether the phase has a point mass at the left/lower boundary

rpm::Bool: whether the phase has a point mass at the right/upper boundary

DiscretisedFluidQueues.PhaseSet — Type

const PhaseSet = Array{Phase,1}

A container for phases.

DiscretisedFluidQueues.PhaseSet — Method

PhaseSet(c::Array{Float64, 1}[, m::Array{Int, 1}, lpm::BitArray, rpm::BitArray])

Construct an Array of phases.

Arguments:

c: Array of rates

m: (optional) array of memberships, see Phase()

lpm: (optional) array of left point mass indicators, see Phase()

rpm: (optional) array of right point mass indicators, see Phase()

DiscretisedFluidQueues.SFMDistribution — Type

SFMDistribution{T <: Mesh} <: AbstractArray{Float64, 2}

Representation of the distribution of a DiscretisedFluidQueue.

The field coeffs is a row-vector and indexing a distribution indexes this field. Similarly, arithmetic operations on a distribution act on this vector.

Arguments:

coeffs::Array{Float64, 2}: A row vector of coefficients which encode the distribution of a DiscretisedFluidQueue and can be used to reconstruct a solution.

dq::DiscretisedFluidQueue{T}:

DiscretisedFluidQueues.SFMDistribution — Type

SFMDistribution(pdf::Function, dq::DiscretisedFluidQueue{FVMesh}, fun_evals::Int = 6)

Construct an approximation to pdf as the average of pdf of each cell.

Uses quadrature with fun_evals function evaluations to approximate cell averages.

DiscretisedFluidQueues.SFMDistribution — Method

SFMDistribution(pdf::Function, dq::DiscretisedFluidQueue{DGMesh})

Approximates pdf via polynomials.

DiscretisedFluidQueues.SFMDistribution — Method

SFMDistribution(pdf::Function, dq::DiscretisedFluidQueue{FRAPMesh}; fun_evals = 100)

Return the appropriate initial condition to approximate the initial distribution pdf for the numerical discretisation scheme defined by the dq DiscretisedFluidQueue.

i.e. for each cell compute the expected orbit position $∫pdf(x,i) ⋅ a(x) dx$ where the integral is over each cell and $a(x) = a exp(S(x-yₖ))$ if the membership of i is -1 and $a(x) = a exp(S(yₖ₊₁-x))$ if the membership of i is 1 where yₖ and yₖ₊₁ are the left and right cell edges, respectively, and a and S are defined by dq.mesh.me are parameters of a MatrixExponential.

DiscretisedFluidQueues.SFMDistribution — Method

SFMDistribution(pdf::Function, dq::DiscretisedFluidQueue{<:Mesh}, fun_evals::Int = 6)

Return a discretised version of the pdf. The method of discretisation depends on the discretisation method of dq.

Arguments:

pdf::Function: a function f(x::Float64,i::Int) where f(x,i)dx=P(X(0)∈dx,φ(0)=i) is the initial distribution of a fluid queue.

dqDiscretisedFluidQueue{<:Mesh}:

fun_evals: the number of function evaluations used to approximate f(x,i)

DiscretisedFluidQueues.SFMDistribution — Method

SFMDistribution(dq::DiscretisedFluidQueue{T})

A blank initialiser for a fluid queue distribution a distribution with coeffs=zeros.

DiscretisedFluidQueues.Simulation — Type

Simulation

A container to hold simulations.

All arguments are arrays of the same ength + a model from which the simulations came

Arguments:

t::Array{Float64, 1}:

φ::Array{Int, 1}:

X::Array{Float64, 1}:

n::Array{Int, 1}:

model::Model:

Base.size — Method

size(B::LazyGenerator)

DiscretisedFluidQueues.UpdateXt — Method

Returns $X(t+S) = min(max(X(t) + cᵢS,0),U)$ where $U$ is some upper bound on the process.

UpdateXt(
    model::Model,
    SFM0::NamedTuple,
    S::Real,
    lwr::Float64,
    upr::Float64,
)

Arguments

model: a Model object

SFM0::NamedTuple containing at least the keys :X giving the value of $X(t)$ at the current time, and :φ giving the value of $φ(t)$ at the current time.

S::Real: an elapsed amount of time to evaluate $X$ at, i.e. $X(t+S)$.

lwr::Float64: lower regulated boundary for the fluid level

upr::Float64: upper regulated boundary for the fluid level

DiscretisedFluidQueues._model_dicts — Method

_model_dicts(model::Model) 

input: a Model object

outputs: - SDict: a dictionary with keys "+","-","0","bullet" and corresponding values findall(model.C .> 0), findall(model.C .< 0), findall(model.C .== 0), findall(model.C .!= 0), respectively. - TDict: a dictionary of submatrices of T with keys "ℓm" with $ℓ,m∈{+,-,0,bullet}$ and corresponding values model.T[S[ℓ],S[m]].

DiscretisedFluidQueues._sum_cells_left — Method

_sum_cells_left(d::SFMDistribution, i::Int, cell_idx::Int, dq::DiscretisedFluidQueue)

Add up all the probability mass in phase i in the cells to the left of cell_idx.

DiscretisedFluidQueues.basis — Method

basis(mesh::DGMesh)

Returns mesh.basis; either "lagrange" or "legendre"

DiscretisedFluidQueues.basis — Method

basis(mesh::FRAPMesh)

Constant ""

DiscretisedFluidQueues.basis — Method

basis(mesh::FVMesh)

Constant ""

DiscretisedFluidQueues.gauss_lobatto_points — Method

gauss_lobatto_points(a::Float64, b::Float64, n_evals::Int)

Return array containing n_evals gauss lobatto points in the interval [a,b].

DiscretisedFluidQueues.gauss_lobatto_quadrature — Method

gauss_lobatto_quadrature(fun::Function, a::Float64, b::Float64, n_evals::Int)

Compute a quadrature approximation of fun on the interval [a,b] with n_evals function evaluations using the gauss_lobatto_points

DiscretisedFluidQueues.gauss_lobatto_weights — Method

gauss_lobatto_weights(a::Float64, b::Float64, n_evals::Int)

Return array containing n_evals weights of the polynomials associated with the gauss lobatto points in the interval [a,b]. i.e. the itegrals over the interval [a,b] of the lagrange polynomials defined by the nodes given by nodes = gauss_lobatto_points(a,b,n_evals)

DiscretisedFluidQueues.lagrange_interpolation — Method

lagrange_interpolation(fun::Function, a::Float64, b::Float64, n_evals::Int)

Return a polynomial approximation to fun of order n_evals-1 on the interval [a,b] using the gauss_lobatto_points as the nodes of the lagrange polynomials.

DiscretisedFluidQueues.lagrange_polynomials — Method

lagrange_polynomials(nodes::Array{Float64, 1}, evalPt::Float64)

Evaluate the lagrange polynomials defied by nodes at the point evalPt.

DiscretisedFluidQueues.local_dg_operators — Method

local_dg_operators(mesh::DGMesh; v::Bool = false)

Construct the block matrices and vectors which define the discretised generator (i.e. the blocks, and boundary_flux in LazyGenerator).

DiscretisedFluidQueues.vandermonde — Method

Construct a generalised vandermonde matrix.

vandermonde( nBases::Int)

Note: requires Jacobi package Pkg.add("Jacobi")

Arguments

nBases::Int: the degree of the basis

Output

a tuple with keys

:V::Array{Float64,2}: where :V[:,i] contains the values of the ith legendre polynomial evaluate at the GLL nodes.

:inv: the inverse of :V

:D::Array{Float64,2}: where V.D[:,i] contains the values of the derivative of the ith legendre polynomial evaluate at the GLL nodes.