Generators

DiscretisedFluidQueues.Generator — Type

Generator

Abstract type representing a discretised infinitesimal generator of a FLuidQueue. Behaves much like a square matrix.

DiscretisedFluidQueues.LazyGenerator — Type

LazyGenerator{<:Mesh} <: Generator{<:Mesh}

A lazy representation of a block matrix with is a generator of a DiscretisedFluidQueue.

Lower memory requirements than FullGenerator but aritmetic operations and indexing may be slower.

Arguments:

dq::DiscretisedFluidQueue:

blocks::Tuple{Array{Float64, 2}, Array{Float64, 2}, Array{Float64, 2}, Array{Float64, 2}}: Block matrices describing the flow of mass within and between cells. blocks[1] is the lower diagonal block describing the flow of mass from cell k+1 to cell (k for phases with negative rate only). blocks[2] (blocks[3]) is the diagonal block describing the flow of mass within a cell for a phase with positive (negative) rate. blocks[4] is the upper diagonal block describing the flow of mass from cell k to k+1 (for phases with positive rate only).

boundary_flux::BoundaryFlux: A named tuple structure such that - boundary_flux.lower.in: describes flow of density into lower boundary - boundary_flux.lower.out: describes flow of density out of lower boundary - boundary_flux.upper.in: describes flow of density into upper boundary - boundary_flux.upper.out: describes flow of density out of upper boundary

D::Union{Array{Float64, 2}, LinearAlgebra.Diagonal{Bool, Array{Bool, 1}}}: An array describing how the flow of density changes when the phase process jumps between phases with different memberships. This is the identity for FV and DG schemes.

DiscretisedFluidQueues.FullGenerator — Type

FullGenerator{T} <: Generator{T}

An Matrix or SparseMatrixCSC representation of the generator of a generator of a DiscretisedFluidQueue.

Higher memory requirements than LazyGenerator (as blocks are duplicated) but much faster matrix arithmetic.

DiscretisedFluidQueues.build_lazy_generator — Function

build_lazy_generator(dq::DiscretisedFluidQueue; v::Bool = false)

Build a lazy representation of the generator of a discretised fluid queue.

DiscretisedFluidQueues.build_full_generator — Function

build_full_generator(dq::DiscretisedFluidQueue; v::Bool = false)

Returns a SparseMatrixCSC generator of a DiscretisedFluidQueue.

build_full_generator(lzB::LazyGenerator)

Returns a SparseMatrixCSC representation of a LazyGenerator.

DiscretisedFluidQueues.static_generator — Function

(Experimantal) convert the block matrices of the lazy generator to static arrays

Base.size — Function

size(B::LazyGenerator)