Models and Discretisations

Model-related Types

DiscretisedFluidQueues.Phase — Type

Phase

Represents a phase of the fluid queue.

Arguments:

c::Float64: the rate of change of the fluid X(t) associated with the phase

m::Int64: either -1 or 1, membership of the phase with either the set of positive or negative phases. Default is positive, i.e. 1. Only relevant for FRAPApproximation.

lpm::Bool: whether the phase has a point mass at the left/lower boundary

rpm::Bool: whether the phase has a point mass at the right/upper boundary

DiscretisedFluidQueues.PhaseSet — Type

const PhaseSet = Array{Phase,1}

A container for phases.

DiscretisedFluidQueues.Model — Type

Model

Abstract type represnting a FluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.DiscretisedFluidQueue — Type

DiscretisedFluidQueue{T <: Mesh}

A container to hold a FluidQueue and a discretisation scheme (a mesh). Parametric type depending on the type of discretisation scheme (<:Mesh)

Arguments:

model::Model: a FluidQueue

mesh::T: either a DGMesh, FVMesh or FRAPMesh specifying the discretisation scheme for the continuous level of the fluid queue, X(t).

DiscretisedFluidQueues.FluidQueue — Type

FluidQueue <: Model

Constructor for a fluid queue model.

Arguments:

T::Array{<:Real, 2}: Generator of the phase process

S::PhaseSet: An array of phases describing the evolution of the fluid level in each phase.

Mesh (discretisation)-related Types

DiscretisedFluidQueues.Mesh — Type

Mesh

Abstract type representing a discretisation mesh for a numerical scheme.

DiscretisedFluidQueues.DGMesh — Type

DGMesh <: Mesh

A structure representing a discretisation scheme to be used for a DiscretisedFluidQueue.

Arguments:

nodes::Array{<:Real, 1}: The edges of the cells.

n_bases::Int: The number of basis functions used to represent the solution on each cell

DiscretisedFluidQueues.FRAPMesh — Type

FRAPMesh <: Mesh

A QBD-RAP (FRAM) discretisation scheme for a DiscretisedFluidQueue.

Arguments

nodes: the cell edges

me: the MatrixExponential used to approximate model the fluid queue on each cell.

DiscretisedFluidQueues.FVMesh — Type

FVMesh <: Mesh

A finite volume discretisation scheme for a DiscretisedFluidQueue.

Arguments

nodes: the cell edges (the nodes are at the center of the cell)

order: the order of the polynomial interpolation used to approximate the flux at the

cell edges.

Model-related Methods

DiscretisedFluidQueues.n_phases — Function

n_phases(S::PhaseSet)

Return the number of phases from a PhaseSet, FluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.rates — Function

rates(S::PhaseSet, i::Int)

Return the rate of the phase from a PhaseSet, FluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.membership — Function

membership(S::PhaseSet, i::Int)

Return the membership of phases from a PhaseSet

DiscretisedFluidQueues.N₋ — Function

N₋(S::PhaseSet)

Return the number of left/lower point masses of a PhaseSet, FluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.N₊ — Function

N₊(S::PhaseSet) = begin

Return the number of right/upper point masses of a PhaseSet, FluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.augment_model — Function

augment_model(model::FluidQueue)

Given a FluidQueue, return a FluidQueue with twice as many phases with rate 0, one set associated with m=1 phases and one associated with m=-1 phases.

Mesh (discretisation)-related Methods

DiscretisedFluidQueues.cell_nodes — Function

cell_nodes(mesh::DGMesh)

The positions of the GLJ nodes within each cell of a mesh.

cell_nodes(mesh::FVMesh)

The cell centres

cell_nodes(mesh::FRAPMesh)

The cell centre

DiscretisedFluidQueues.n_intervals — Function

n_intervals(m)

Total number of cells for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.Δ — Function

Δ(m)

The width of each cell for a Mesh or DiscretisedFluidQueue.

Δ(m,k::Int)

The width of cell k for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.n_bases_per_phase — Function

n_bases_per_phase(m)

Total number of basis functions used to represent the fluid queue for each phase of a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.n_bases_per_cell — Function

n_bases_per_cell(m) =

Return the number of basis functions on each cell of a given DiscretisedFluidQueue or Mesh.

n_bases_per_cell(mesh::DGMesh)

Number of bases in a cell

n_bases_per_cell(mesh::FVMesh)

Constant 1

n_bases_per_cell(mesh::FRAPMesh)

Number of bases in a cell

DiscretisedFluidQueues.total_n_bases — Function

total_n_bases(m)

Return the number of basis functions of a given DiscretisedFluidQueue or Mesh.