ME Tools

Types

DiscretisedFluidQueues.AbstractMatrixExponential — Type

AbstractMatrixExponential

DiscretisedFluidQueues.ConcentratedMatrixExponential — Type

Similar to MatrixExponential but has a neater implementation of orbit() due to nice structure.

DiscretisedFluidQueues.MatrixExponential — Type

MatrixExponential constructor method

MatrixExponential(
    a::Array{Float64,2},
    S::Array{Float64,2},
    s::Array{Float64,1},
    D::Array{Float64}=[0],
)

Inputs:

a a 1 by p Array of reals

S a p by p Array of reals

s a p by 1 Array of reals

D an optional argument, if empty then me.D is the identity, else is a p by p matrix

Throws an error if the dimensions are inconsistent.

DiscretisedFluidQueues.cme_params — Constant

cme_params

A Dict storing parameters of concentrated matrix exponential distributions. See https://github.com/ghorvath78/iltcme.

Methods

DiscretisedFluidQueues.build_me — Function

build_me(params; mean::Real = 1)

Return a ConcentratedMatrixExponential as defined by the dictionary params. i.e. see CMEParams[3] for an order 3 CME.

DiscretisedFluidQueues.pdf — Method

pdf([a::Array{Float64, 2},] me::AbstractMatrixExponential)

Return a the probability density function of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)

DiscretisedFluidQueues.ccdf — Method

ccdf(a::Array{Float64, 2}, me::AbstractMatrixExponential)

Return a the complimentary cumulative distribution function (1-cdf) of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)

DiscretisedFluidQueues.cdf — Method

cdf(a::Array{Float64, 2}, me::AbstractMatrixExponential) = begin

Return a the cumulative distribution function of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)