Mesh (discretisation)-related Types

DiscretisedFluidQueues.Mesh — Type

Mesh

Abstract type representing a discretisation mesh for a numerical scheme. Possible subtypes are DGMesh, FRAPMesh, FVMesh

DiscretisedFluidQueues.DGMesh — Type

DGMesh{T} <: Mesh{T}

A structure representing a discretisation scheme to be used for a DiscretisedFluidQueue.

Arguments:

nodes::AbstractArray{Float64, 1}: The edges of the cells.

n_bases::Int: The number of basis functions used to represent the solution on each cell

DiscretisedFluidQueues.FRAPMesh — Type

FRAPMesh{T} <: Mesh{T}

A QBD-RAP discretisation scheme for a DiscretisedFluidQueue.

Arguments

nodes: the cell edges

me: the MatrixExponential used to approximate model the fluid queue on each cell.

DiscretisedFluidQueues.FVMesh — Type

FVMesh{T} <: Mesh{T}

A finite volume discretisation scheme for a DiscretisedFluidQueue.

Arguments

nodes: the cell edges (the nodes are at the center of the cell)

order: the order of the polynomial interpolation used to approximate the flux at the

cell edges.

Mesh (discretisation)-related Methods

DiscretisedFluidQueues.cell_nodes — Function

cell_nodes(mesh::DGMesh)

The positions of the GLJ nodes within each cell of a mesh.

cell_nodes(mesh::FVMesh)

The cell centres

cell_nodes(mesh::FRAPMesh)

The cell centre

DiscretisedFluidQueues.n_intervals — Function

n_intervals(m)

Total number of cells for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.Δ — Function

Δ(m)

The width of each cell for a Mesh or DiscretisedFluidQueue.

Δ(m,k::Int)

The width of cell k for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.n_bases_per_phase — Function

n_bases_per_phase(m)

Total number of basis functions used to represent the fluid queue for each phase of a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.n_bases_per_cell — Function

n_bases_per_cell(m) =

Return the number of basis functions on each cell of a given DiscretisedFluidQueue or Mesh.

n_bases_per_cell(mesh::DGMesh)

Number of bases in a cell

n_bases_per_cell(mesh::FVMesh)

Constant=1

n_bases_per_cell(mesh::FRAPMesh)

Number of bases in a cell

DiscretisedFluidQueues.total_n_bases — Function

total_n_bases(m)

Return the number of basis functions of a given DiscretisedFluidQueue or Mesh.