Methods

DiscretisedFluidQueues.N₊ — Method

N₊(S::PhaseSet) = begin

Return the number of right/upper point masses of a PhaseSet, FluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.N₋ — Method

N₋(S::PhaseSet)

Return the number of left/lower point masses of a PhaseSet, FluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.augment_model — Method

augment_model(model::FluidQueue)

Given a FluidQueue, return a FluidQueue with twice as many phases with rate 0, one set associated with m=1 phases and one associated with m=-1 phases.

DiscretisedFluidQueues.build_full_generator — Method

build_full_generator(dq::DiscretisedFluidQueue; v::Bool = false)

Returns a SparseMatrixCSC generator of a DiscretisedFluidQueue.

DiscretisedFluidQueues.build_full_generator — Method

build_full_generator(lzB::LazyGenerator)

Returns a SparseMatrixCSC representation of a LazyGenerator.

DiscretisedFluidQueues.build_lazy_generator — Method

build_lazy_generator(dq::DiscretisedFluidQueue; v::Bool = false)

Build a lazy representation of the generator of a discretised fluid queue.

DiscretisedFluidQueues.build_me — Method

build_me(params; mean::Real = 1)

Return a ConcentratedMatrixExponential as defined by the dictionary params. i.e. see CMEParams[3] for an order 3 CME.

DiscretisedFluidQueues.ccdf — Method

ccdf(a::Array{Float64, 2}, me::AbstractMatrixExponential, x::Real)

Return a the complimentary cumulative distribution function (1-cdf) of a matrix exponential evaluated at x.

DiscretisedFluidQueues.ccdf — Method

ccdf(a::Array{Float64, 2}, me::AbstractMatrixExponential)

Return a the complimentary cumulative distribution function (1-cdf) of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)

DiscretisedFluidQueues.cdf — Method

cdf(a::Array{Float64, 2}, me::AbstractMatrixExponential, x::Real)

Return a the cumulative distribution function of a matrix exponential evaluated at x.

DiscretisedFluidQueues.cdf — Method

cdf(a::Array{Float64, 2}, me::AbstractMatrixExponential) = begin

Return a the cumulative distribution function of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)

DiscretisedFluidQueues.cdf — Method

cdf(d::SFMDistribution, x, i)

Evaluate cdf(d::SFMDistribution) at (x,i).

DiscretisedFluidQueues.cdf — Method

cdf(d::SFMDistribution)

Return a function of two variables (x::Float64,i::Int) which is the cumulative distribution distribution function defined by d. Approximates the distribution function of a fluid queue.

DiscretisedFluidQueues.cdf — Method

cdf(s::Simulation)

Return a empirical cdf of the fluid queue given a simulation.

Output is a function of two variables (x,i) and gives the empirical distribution function $P(X(τ)≤x,φ(τ)=i)$

DiscretisedFluidQueues.cell_nodes — Method

cell_nodes(mesh::DGMesh)

The positions of the GLJ nodes within each cell of a mesh.

DiscretisedFluidQueues.cell_nodes — Method

cell_nodes(mesh::FRAPMesh)

The cell centre

DiscretisedFluidQueues.cell_nodes — Method

cell_nodes(mesh::FVMesh)

The cell centres

DiscretisedFluidQueues.first_exit_x — Method

Constructs the StoppingTime which is the first exit of the process $X(t)$ from the interval $[u,v]$.

first_exit_x( u::Real, v::Real)

Arguments

u: a lower boundary

v: an upper boundary

Output

first_exit_xFun: a function with one methods

`firstexitxFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.

DiscretisedFluidQueues.fixed_time — Method

Constructs the StoppingTime $1(t>T)$

fixed_time( T::Real)

Arguments

T: a time at which to stop the process

Output

fixed_timeFun: a function with one methods

`fixed_timeFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.

DiscretisedFluidQueues.integrate_time — Function

Given x0 apprximate x0 exp(Dy).

integrate_time(x0::AbstractArray{<:Real, 2}, D::Generator, y::Real, h::TimeIntegrationScheme = Euler(y ./ 1000))

Arguments

x0: An initial row vector

D: A square matrix

y: time to integrate up to

h: TimeIntegrationScheme.

Output

DiscretisedFluidQueues.interior_point_mass — Method

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{DGMesh})

Constructs a polynomial approximation to the point mass at (x,i)

DiscretisedFluidQueues.interior_point_mass — Method

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{FRAPMesh})

Return the appropriate initial condition to approximate the initial distribution which is a point mass at (x,i) for the numerical discretisation scheme defined by the dq DiscretisedFluidQueue.

i.e. Compute the orbit position $a(x) = a exp(S(x-yₖ))$ if the membership of i is -1 and $a(x) = a exp(S(yₖ₊₁-x))$ if the membership of i is 1 where yₖ and yₖ₊₁ are the left and right cell edges and x∈[yₖ,yₖ₊₁], respectively, and a and S are defined by dq.mesh.me are parameters of a MatrixExponential.

DiscretisedFluidQueues.interior_point_mass — Method

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{FVMesh})

Construct an approximation to a point mass at (x,i). Basically, just puts mass 1 in the cell containing (x,i).

DiscretisedFluidQueues.interior_point_mass — Method

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{<:Mesh})

Construct an approximation to a distribution of fluid queue with a point mass at (x,i). Returns a SFMDistribution. The method of approximation is defined by the type of <:Mesh.

Arguments:

x::Float64: The position of the point mass

i::Int: The phase of the point mass

dq::DiscretisedFluidQueue:

DiscretisedFluidQueues.left_point_mass — Method

left_point_mass(i::Int, dq::DiscretisedFluidQueue)

Construct a SFMDistribution with a point mass at the left boundary of dq in phase i and 0 elsewhere.

DiscretisedFluidQueues.membership — Method

membership(S::PhaseSet, i::Int)

Return the membership of phases from a PhaseSet

DiscretisedFluidQueues.n_bases_per_cell — Method

n_bases_per_cell(mesh::DGMesh)

Number of bases in a cell

DiscretisedFluidQueues.n_bases_per_cell — Method

n_bases_per_cell(m) =

Return the number of basis functions on each cell of a given DiscretisedFluidQueue or Mesh.

DiscretisedFluidQueues.n_bases_per_cell — Method

n_bases_per_cell(mesh::FRAPMesh)

Number of bases in a cell

DiscretisedFluidQueues.n_bases_per_cell — Method

n_bases_per_cell(mesh::FVMesh)

Constant 1

DiscretisedFluidQueues.n_bases_per_phase — Method

n_bases_per_phase(m)

Total number of basis functions used to represent the fluid queue for each phase of a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.n_intervals — Method

n_intervals(m)

Total number of cells for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.n_jumps — Method

Constructs the StoppingTime $1(N(t)>n)$ where $N(t)$ is the number of jumps of $φ$ by time $t$.

n_jumps( N::Int)

Arguments

N: a desired number of jumps

Output

n_jumpsFun: a function with one methods

`n_jumpsFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.

DiscretisedFluidQueues.n_phases — Method

n_phases(S::PhaseSet)

Return the number of phases from a PhaseSet, FluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.pdf — Method

pdf(a::Array{Float64, 2}, me::AbstractMatrixExponential, x::Real)

Return a the probability density function of a matrix exponential evaluated at x.

DiscretisedFluidQueues.pdf — Method

pdf([a::Array{Float64, 2},] me::AbstractMatrixExponential)

Return a the probability density function of a matrix exponential as a function.

If a is not specified then the initial vector of the me is me.a

Arguments:

a: a row vector with length size(me.S,1)

DiscretisedFluidQueues.pdf — Method

pdf(d::SFMDistribution, x, i)

Evaluate pdf(d::SFMDistribution) at (x,i).

DiscretisedFluidQueues.pdf — Method

pdf(d::SFMDistribution)

Return a function of two variables (x::Float64,i::Int) which is the probability distribution function defined by d. Approximates the density function of a fluid queue.

DiscretisedFluidQueues.phases — Method

phases(S::PhaseSet)

Return the iterator 1:n_phases(S), for a PhaseSet, FluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.psi_fun_x — Method

Construct and evaluate $Ψ(s)$ for a triditional SFM.

Uses newtons method to solve the Ricatti equation $D⁺⁻(s) + Ψ(s)D⁻⁺(s)Ψ(s) + Ψ(s)D⁻⁻(s) + D⁺⁺(s)Ψ(s) = 0.$

psi_fun_x( model::Model; s = 0, MaxIters = 1000, err = 1e-8)

Arguments

model: a Model object

s::Real: a value to evaluate the LST at

MaxIters::Int: the maximum number of iterations of newtons method

err::Float64: an error tolerance for terminating newtons method. Terminates when max(Ψ_{n} - Ψ{n-1}) .< eps.

Output

Ψ(s)::Array{Float64,2} the matrix $Ψ$

DiscretisedFluidQueues.rates — Method

rates(S::PhaseSet, i::Int)

Return the rate of the phase from a PhaseSet, FluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.right_point_mass — Method

right_point_mass(i::Int, dq::DiscretisedFluidQueue)

Construct a SFMDistribution with a point mass at the left boundary of dq in phase i and 0 elsewhere.

DiscretisedFluidQueues.simulate — Function

Simulates a SFM defined by Model until the StoppingTime has occured, given the InitialCondition on (φ(0),X(0)).

simulate(
    model::Model,
    lwr::Float64, upr::Float64,
    StoppingTime::Function,
    InitCondition::NamedTuple{(:φ, :X)},
)

Arguments

model: A Model object

lwr and upr upper an lower regulated boundaries for the fluid level.

StoppingTime: A function which takes the value of the process at the current time and at the time of the last jump of the phase process, as well as the model object. i.e. StoppingTime(;model,SFM,SFM0,lwer,upr) where SFM and SFM0 are tuples with keys (:t::Float64, :φ::Int, :X::Float64, :n::Int) which are the value of the SFM at the current time, and time of the previous jump of the phase process, repsectively. The StoppingTime must return a NamedTuple{(:Ind, :SFM)} type where :Ind is a :Bool value stating whether the stopping time has occured or not and :SFM is a tuple in the same form as the input SFM but which contains the value of the SFM at the stopping time.

InitCondition: NamedTuple with keys (:φ, :X), InitCondition.φ is a vector of length M of initial states for the phase, and InitCondition.X is a vector of length M of initial states for the level. M is the number of simulations to be done.

Output

A Simulation object.

DiscretisedFluidQueues.stationary_distribution_x — Method

Construct the stationary distribution of the SFM

stationary_distribution_x( model::Model, Ψ::Array, ξ::Array)

Arguments

model: a Model object

Ψ: an array as output from psi_fun_x

ξ: an array as returned from xi_x

Output

pₓ::Array{Float64,2}: the point masses of the SFM

πₓ(x) a function with two methods

πₓ(x::Real): for scalar inputs, returns the stationary density evaluated at x in all phases.

πₓ(x::Array): for array inputs, returns an array with the same shape as is output by Coeff2Dist.

K::Array{Float64,2}: the matrix in the exponential of the density.

DiscretisedFluidQueues.total_n_bases — Method

total_n_bases(m)

Return the number of basis functions of a given DiscretisedFluidQueue or Mesh.

DiscretisedFluidQueues.xi_x — Method

Construct the vector $ξ$ containing the distribution of the phase at the time when $X(t)$ first hits 0.

xi_x( model::Model, Ψ::Array)

Arguments

model: a Model object

Ψ: an array as output from psi_fun_x

Output

the vector ξ

DiscretisedFluidQueues.Δ — Method

Δ(m,k::Int)

The width of cell k for a Mesh or DiscretisedFluidQueue.

DiscretisedFluidQueues.Δ — Method

Δ(m)

The width of each cell for a Mesh or DiscretisedFluidQueue.

Base.size — Method

size(B::LazyGenerator)

DiscretisedFluidQueues.UpdateXt — Method

Returns $X(t+S) = min(max(X(t) + cᵢS,0),U)$ where $U$ is some upper bound on the process.

UpdateXt(
    model::Model,
    SFM0::NamedTuple,
    S::Real,
    lwr::Float64,
    upr::Float64,
)

Arguments

model: a Model object

SFM0::NamedTuple containing at least the keys :X giving the value of $X(t)$ at the current time, and :φ giving the value of $φ(t)$ at the current time.

S::Real: an elapsed amount of time to evaluate $X$ at, i.e. $X(t+S)$.

lwr::Float64: lower regulated boundary for the fluid level

upr::Float64: upper regulated boundary for the fluid level

DiscretisedFluidQueues._model_dicts — Method

_model_dicts(model::Model) 

input: a Model object

outputs: - SDict: a dictionary with keys "+","-","0","bullet" and corresponding values findall(model.C .> 0), findall(model.C .< 0), findall(model.C .== 0), findall(model.C .!= 0), respectively. - TDict: a dictionary of submatrices of T with keys "ℓm" with $ℓ,m∈{+,-,0,bullet}$ and corresponding values model.T[S[ℓ],S[m]].

DiscretisedFluidQueues._sum_cells_left — Method

_sum_cells_left(d::SFMDistribution, i::Int, cell_idx::Int, dq::DiscretisedFluidQueue)

Add up all the probability mass in phase i in the cells to the left of cell_idx.

DiscretisedFluidQueues.basis — Method

basis(mesh::DGMesh)

Returns mesh.basis; either "lagrange" or "legendre"

DiscretisedFluidQueues.basis — Method

basis(mesh::FRAPMesh)

Constant ""

DiscretisedFluidQueues.basis — Method

basis(mesh::FVMesh)

Constant ""

DiscretisedFluidQueues.gauss_lobatto_points — Method

gauss_lobatto_points(a::Float64, b::Float64, n_evals::Int)

Return array containing n_evals gauss lobatto points in the interval [a,b].

DiscretisedFluidQueues.gauss_lobatto_quadrature — Method

gauss_lobatto_quadrature(fun::Function, a::Float64, b::Float64, n_evals::Int)

Compute a quadrature approximation of fun on the interval [a,b] with n_evals function evaluations using the gauss_lobatto_points

DiscretisedFluidQueues.gauss_lobatto_weights — Method

gauss_lobatto_weights(a::Float64, b::Float64, n_evals::Int)

Return array containing n_evals weights of the polynomials associated with the gauss lobatto points in the interval [a,b]. i.e. the itegrals over the interval [a,b] of the lagrange polynomials defined by the nodes given by nodes = gauss_lobatto_points(a,b,n_evals)

DiscretisedFluidQueues.lagrange_interpolation — Method

lagrange_interpolation(fun::Function, a::Float64, b::Float64, n_evals::Int)

Return a polynomial approximation to fun of order n_evals-1 on the interval [a,b] using the gauss_lobatto_points as the nodes of the lagrange polynomials.

DiscretisedFluidQueues.lagrange_polynomials — Method

lagrange_polynomials(nodes::Array{Float64, 1}, evalPt::Float64)

Evaluate the lagrange polynomials defied by nodes at the point evalPt.

DiscretisedFluidQueues.local_dg_operators — Method

local_dg_operators(mesh::DGMesh; v::Bool = false)

Construct the block matrices and vectors which define the discretised generator (i.e. the blocks, and boundary_flux in LazyGenerator).

DiscretisedFluidQueues.vandermonde — Method

Construct a generalised vandermonde matrix.

vandermonde( nBases::Int)

Note: requires Jacobi package Pkg.add("Jacobi")

Arguments

nBases::Int: the degree of the basis

Output

a tuple with keys

:V::Array{Float64,2}: where :V[:,i] contains the values of the ith legendre polynomial evaluate at the GLL nodes.

:inv: the inverse of :V

:D::Array{Float64,2}: where V.D[:,i] contains the values of the derivative of the ith legendre polynomial evaluate at the GLL nodes.