Models and Discretisations

Model-related Types

DiscretisedFluidQueues.Phase — Type

Phase

Represents a phase of the fluid queue.

Arguments:

c::Float64: the rate of change of the fluid X(t) associated with the phase

m::Int64: either -1 or 1, membership of the phase with either the set of positive or negative phases. Default is positive, i.e. 1. Only relevant for FRAPApproximation.

lpm::Bool: whether the phase has a point mass at the left/lower boundary

rpm::Bool: whether the phase has a point mass at the right/upper boundary

DiscretisedFluidQueues.PhaseSet — Type

const PhaseSet = Array{Phase,1}

A container for phases.

DiscretisedFluidQueues.Model — Type

Model

Abstract type represnting a BoundedFluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.DiscretisedFluidQueue — Type

DiscretisedFluidQueue{T <: Mesh}

A container to hold a BoundedFluidQueue and a discretisation scheme (a mesh). Parametric type depending on the type of discretisation scheme (<:Mesh)

Arguments:

model::Model: a BoundedFluidQueue

mesh::T: either a DGMesh, FVMesh or FRAPMesh specifying the discretisation scheme for the continuous level of the fluid queue, X(t).

DiscretisedFluidQueues.BoundedFluidQueue — Type

BoundedFluidQueue <: Model

Constructor for a fluid queue model.

Arguments:

T::Array{Float64, 2}: Generator of the phase process

S::PhaseSet: An array of phases describing the evolution of the fluid level in each phase.

P_lwr::Array{Float64,2}: Probabilities of phase change upon hitting the lower boundary. A |S₋| by |S| array.

P_upr::Array{Float64,2}: Probabilities of phase change upon hitting the upper boundary. A |S₊| by |S| array.

b::Float64: the upper boundary of the fluid level (the lower boundary is 0 by default).

Model-related Methods

DiscretisedFluidQueues.n_phases — Function

n_phases(S::PhaseSet)

Return the number of phases from a PhaseSet, BoundedFluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.rates — Function

rates(S::PhaseSet, i::Int)

Return the rate of the phase from a PhaseSet, BoundedFluidQueue, or DiscretisedFluidQueue

DiscretisedFluidQueues.membership — Function

membership(S::PhaseSet, i::Int)

Return the membership of phases from a PhaseSet

DiscretisedFluidQueues.N₋ — Function

N₋(S::PhaseSet)

Return the number of left/lower point masses of a PhaseSet, BoundedFluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.N₊ — Function

N₊(S::PhaseSet) = begin

Return the number of right/upper point masses of a PhaseSet, BoundedFluidQueue or DiscretisedFluidQueue

DiscretisedFluidQueues.augment_model — Function

augment_model(model::Model)

Given a BoundedFluidQueue, return a BoundedFluidQueue with twice as many phases with rate 0, one set associated with m=1 phases and one associated with m=-1 phases.