Polynomial interpolation and tools

DG and FV methods rely on polynomial approximations.

DiscretisedFluidQueues.lagrange_polynomials — Function

lagrange_polynomials(nodes::Array{Float64, 1}, evalPt::Float64)

Evaluate the lagrange interpolating polynomials defied by nodes at the point evalPt.

DiscretisedFluidQueues.gauss_lobatto_points — Function

gauss_lobatto_points(a::Float64, b::Float64, n_evals::Int)

Return array containing n_evals gauss lobatto points in the interval [a,b].

DiscretisedFluidQueues.gauss_lobatto_weights — Function

gauss_lobatto_weights(a::Float64, b::Float64, n_evals::Int)

Return array containing n_evals weights of the polynomials associated with the gauss lobatto points in the interval [a,b]. i.e. the itegrals over the interval [a,b] of the lagrange polynomials defined by the nodes given by nodes = gauss_lobatto_points(a,b,n_evals)

DiscretisedFluidQueues.lagrange_interpolation — Function

lagrange_interpolation(fun::Function, a::Float64, b::Float64, n_evals::Int)

Return a polynomial approximation to fun of order n_evals-1 on the interval [a,b] using the gauss_lobatto_points as the nodes of the lagrange polynomials.

DiscretisedFluidQueues.gauss_lobatto_quadrature — Function

gauss_lobatto_quadrature(fun::Function, a::Float64, b::Float64, n_evals::Int)

Compute a quadrature approximation of fun on the interval [a,b] with n_evals function evaluations using the gauss_lobatto_points

gauss_lobatto_quadrature(fun::Function, nodes::Array{Float64,1})

Compute a quadrature approximation of fun on the interval nodes[1] to nodes[end] evaluating the function fun at nodes

DiscretisedFluidQueues.vandermonde — Function

Construct a generalised vandermonde matrix.

vandermonde( nBases::Int)

Note: requires Jacobi package Pkg.add("Jacobi")

Arguments

nBases::Int: the degree of the basis

Output

a tuple with keys

:V::Array{Float64,2}: where :V[:,i] contains the values of the ith legendre polynomial evaluate at the GLL nodes.

:inv: the inverse of :V

:D::Array{Float64,2}: where V.D[:,i] contains the values of the derivative of the ith legendre polynomial evaluate at the GLL nodes.