SFM operators

Tools to evaluate the stationary distribution of a FluidQueue via matrix-analytic methods.

DiscretisedFluidQueues.psi_fun_x — Function

Construct and evaluate $Ψ(s)$ for a triditional SFM.

Uses newtons method to solve the Ricatti equation $D⁺⁻(s) + Ψ(s)D⁻⁺(s)Ψ(s) + Ψ(s)D⁻⁻(s) + D⁺⁺(s)Ψ(s) = 0.$

psi_fun_x( model::Model; s = 0, MaxIters = 1000, err = 1e-8)

Arguments

model: a Model object

s::Real: a value to evaluate the LST at

MaxIters::Int: the maximum number of iterations of newtons method

err::Float64: an error tolerance for terminating newtons method. Terminates when max(Ψ_{n} - Ψ{n-1}) .< eps.

Output

Ψ(s)::Array{Float64,2} the matrix $Ψ$

DiscretisedFluidQueues.xi_x — Function

Construct the vector $ξ$ containing the distribution of the phase at the time when $X(t)$ first hits 0.

xi_x( model::Model, Ψ::Array)

Arguments

model: a Model object

Ψ: an array as output from psi_fun_x

Output

the vector ξ

DiscretisedFluidQueues.stationary_distribution_x — Function

Construct the stationary distribution of the SFM

stationary_distribution_x( model::Model, Ψ::Array, ξ::Array)

Arguments

model: a Model object

Ψ: an array as output from psi_fun_x

ξ: an array as returned from xi_x

Output

pₓ::Array{Float64,2}: the point masses of the SFM

πₓ(x) a function with two methods

πₓ(x::Real): for scalar inputs, returns the stationary density evaluated at x in all phases.

πₓ(x::Array): for array inputs, returns an array with the same shape as is output by Coeff2Dist.

K::Array{Float64,2}: the matrix in the exponential of the density.