Simulate

Tools to simulate a FluidQueue.

Simulate and evaluate simulations

DiscretisedFluidQueues.Simulation — Type

Simulation

A container to hold simulations.

All arguments are arrays of the same ength + a model from which the simulations came

Arguments:

t::Array{Float64, 1}:

φ::Array{Int, 1}:

X::Array{Float64, 1}:

n::Array{Int, 1}:

model::Model:

DiscretisedFluidQueues.simulate — Function

Simulates a SFM defined by Model until the StoppingTime has occured, given the InitialCondition on (φ(0),X(0)).

simulate(
    model::Model,
    lwr::Float64, upr::Float64,
    StoppingTime::Function,
    InitCondition::NamedTuple{(:φ, :X)},
)

Arguments

model: A Model object

lwr and upr upper an lower regulated boundaries for the fluid level.

StoppingTime: A function which takes the value of the process at the current time and at the time of the last jump of the phase process, as well as the model object. i.e. StoppingTime(;model,SFM,SFM0,lwer,upr) where SFM and SFM0 are tuples with keys (:t::Float64, :φ::Int, :X::Float64, :n::Int) which are the value of the SFM at the current time, and time of the previous jump of the phase process, repsectively. The StoppingTime must return a NamedTuple{(:Ind, :SFM)} type where :Ind is a :Bool value stating whether the stopping time has occured or not and :SFM is a tuple in the same form as the input SFM but which contains the value of the SFM at the stopping time.

InitCondition: NamedTuple with keys (:φ, :X), InitCondition.φ is a vector of length M of initial states for the phase, and InitCondition.X is a vector of length M of initial states for the level. M is the number of simulations to be done.

Output

A Simulation object.

DiscretisedFluidQueues.cdf — Method

cdf(s::Simulation)

Return a empirical cdf of the fluid queue given a simulation.

Output is a function of two variables (x,i) and gives the empirical distribution function $P(X(τ)≤x,φ(τ)=i)$

Predefined stopping times

DiscretisedFluidQueues.fixed_time — Function

Constructs the StoppingTime $1(t>T)$

fixed_time( T::Real)

Arguments

T: a time at which to stop the process

Output

fixed_timeFun: a function with one methods

`fixed_timeFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.

DiscretisedFluidQueues.n_jumps — Function

Constructs the StoppingTime $1(N(t)>n)$ where $N(t)$ is the number of jumps of $φ$ by time $t$.

n_jumps( N::Int)

Arguments

N: a desired number of jumps

Output

n_jumpsFun: a function with one methods

`n_jumpsFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.

DiscretisedFluidQueues.first_exit_x — Function

Constructs the StoppingTime which is the first exit of the process $X(t)$ from the interval $[u,v]$.

first_exit_x( u::Real, v::Real)

Arguments

u: a lower boundary

v: an upper boundary

Output

first_exit_xFun: a function with one methods

`firstexitxFun( model::Model, SFM::NamedTuple{(:t, :φ, :X, :n)}, SFM0::NamedTuple{(:t, :φ, :X, :n)}, lwr::Float64, upr::Float64,

)`: a stopping time for a SFM.