Integrate time

DiscretisedFluidQueues.ExplicitRungeKuttaScheme — Type

ExplicitRungeKuttaScheme(
    step_size::Float64,
    alpha::LinearAlgebra.LowerTriangular{Float64},
    beta::LinearAlgebra.LowerTriangular{Float64},
)

Generic data container for ExplicitRungeKuttaScheme for use in integrate_time.

We are interested in numerically integrating ODEs of the form $a'(t) = a(t) Q$

where $a(t)$ is a vector of coefficients and $Q$ is a matrix, given an initial condition $a(0)$.

In this context, Explicit Runge-Kutta methods with $s$ stages can be written as

$v^{(0)} = a(t),$

$v^{(l)} = \sum_{k=0}^{l-1} \alpha_{l k} v^{(k)} + h\beta_{l k} v^{(k)} Q,\, l = 1,...,s,$

$a(t+h) =v^{(s)},$

where $h$ is the step-size and $\alpha_{l k}$ and $\beta_{l k}$ are parameters which define the scheme.

Arguments

step_size::Float64: the step size of the integration scheme.

alpha::LinearAlgebra.LowerTriangular{Float64}: coefficients as described above

beta::LinearAlgebra.LowerTriangular{Float64}: coefficients as described above

DiscretisedFluidQueues.ForwardEuler — Function

ForwardEuler(step_size::Float64) <: ExplicitRungeKuttaScheme

Defines an Euler integration scheme to be used in integrate_time.

Arguments

step_size::Float64: the step size of the integration scheme.

DiscretisedFluidQueues.Heuns — Function

Heuns(step_size::Float64) <: ExplicitRungeKuttaScheme

Defines Heuns integration scheme to be used in integrate_time.

Arguments

step_size::Float64: the step size of the integration scheme.

DiscretisedFluidQueues.StableRK3 — Function

StableRK3(step_size::Float64) <: ExplicitRungeKuttaScheme

Defines a strong stability preserving Runge-Kutta integration scheme to be used in integrate_time.

Arguments

step_size::Float64: the step size of the integration scheme.

Hesthaven, J. S. & Warburton, T. (2007), Nodal discontinuous Galerkin methods: algorithms, analysis, and applications, Springer Science & Business Media. (Section 5.7)

DiscretisedFluidQueues.StableRK4 — Function

StableRK4(step_size::Float64) = ExplicitRungeKuttaScheme(
        step_size,
        LinearAlgebra.LowerTriangular(_α),
        LinearAlgebra.LowerTriangular(_β)
    )

Defines a strong stability preserving Runge-Kutta integration scheme to be used in integrate_time.

Arguments

step_size::Float64: the step size of the integration scheme.

Raymond J. Spiteri and Steven J. Ruuth. A new class of optimal high-order strong-stability-preserving time discretization methods. SIAM J. Numer. Anal., 40(2):469-491, 2002.

Hesthaven, J. S. & Warburton, T. (2007), Nodal discontinuous Galerkin methods: algorithms, analysis, and applications, Springer Science & Business Media. (Section 5.7)

DiscretisedFluidQueues.integrate_time — Function

Given x0 and D apprximate x0 exp(Dy).

integrate_time(x0::Array{Float64,2}, D::AbstractArray{Float64,2},
    y::Float64, scheme::ExplicitRungeKuttaScheme [; limiter])

Arguments

x0: An initial vector

D: A square matrix

y: time to integrate up to

h: ExplicitRungeKuttaScheme.

limiter: optional named argument spacifying the slope-limiter to used. The limiter function will be applied at every stage and every time-step of the ExplicitRungeKuttaScheme.