Types

DiscretisedFluidQueues.ConcentratedMatrixExponential — Type

Similar to MatrixExponential but has a neater implementation of orbit() due to nice structure.

DiscretisedFluidQueues.DGMesh — Type

DGMesh <: Mesh

A structure representing a discretisation scheme to be used for a DiscretisedFluidQueue.

Arguments:

nodes::Array{<:Real, 1}: The edges of the cells.

n_bases::Int: The number of basis functions used to represent the solution on each cell

DiscretisedFluidQueues.DiscretisedFluidQueue — Type

DiscretisedFluidQueue{T <: Mesh}

A container to hold a FluidQueue and a discretisation scheme (a mesh). Parametric type depending on the type of discretisation scheme (<:Mesh)

Arguments:

model::Model: a FluidQueue

mesh::T: either a DGMesh, FVMesh or FRAPMesh specifying the discretisation scheme for the continuous level of the fluid queue, X(t).

DiscretisedFluidQueues.Euler — Type

Euler <: TimeIntegrationScheme

Defines an Euler integration scheme to be used in integrate_time.

Arguments

step_size::Float64: the step size of the integration scheme.

DiscretisedFluidQueues.FRAPMesh — Type

FRAPMesh <: Mesh

A QBD-RAP (FRAM) discretisation scheme for a DiscretisedFluidQueue.

Arguments

nodes: the cell edges

me: the MatrixExponential used to approximate model the fluid queue on each cell.

DiscretisedFluidQueues.FVMesh — Type

FVMesh <: Mesh

A finite volume discretisation scheme for a DiscretisedFluidQueue.

Arguments

nodes: the cell edges (the nodes are at the center of the cell)

order: the order of the polynomial interpolation used to approximate the flux at the

cell edges.

DiscretisedFluidQueues.FluidQueue — Type

FluidQueue <: Model

Constructor for a fluid queue model.

Arguments:

T::Array{<:Real, 2}: Generator of the phase process

S::PhaseSet: An array of phases describing the evolution of the fluid level in each phase.

DiscretisedFluidQueues.FullGenerator — Type

FullGenerator <: Generator

An Matrix or SparseMatrixCSC representation of the generator of a generator of a DiscretisedFluidQueue.

Higher memory requirements than LazyGenerator (as blocks are duplicated) but much faster matrix arithmetic.

DiscretisedFluidQueues.Generator — Type

Generator

Abstract type representing a discretised infinitesimal generator of a FLuidQueue.

DiscretisedFluidQueues.LazyGenerator — Type

LazyGenerator <: Generator

A lazy representation of a block matrix with is a generator of a DiscretisedFluidQueue.

Lower memory requirements than FullGenerator but aritmetic operations and indexing may be slower.

Arguments:

dq::DiscretisedFluidQueue:

blocks::Tuple{Array{Float64, 2}, Array{Float64, 2}, Array{Float64, 2}, Array{Float64, 2}}: Block matrices describing the flow of mass within and between cells. blocks[1] is the lower diagonal block describing the flow of mass from cell k+1 to cell (k for phases with negative rate only). blocks[2] (blocks[3]) is the diagonal block describing the flow of mass within a cell for a phase with positive (negative) rate. blocks[4] is the upper diagonal block describing the flow of mass from cell k to k+1 (for phases with positive rate only).

boundary_flux::BoundaryFluxTupleType: A named tuple structure such that - boundary_flux.lower.in: describes flow of density into lower boundary - boundary_flux.lower.out: describes flow of density out of lower boundary - boundary_flux.upper.in: describes flow of density into upper boundary - boundary_flux.upper.out: describes flow of density out of upper boundary

D::Union{Array{Float64, 2}, LinearAlgebra.Diagonal{Bool, Array{Bool, 1}}}: An array describing how the flow of density changes when the phase process jumps between phases with different memberships. This is the identity for FV and DG schemes.

DiscretisedFluidQueues.MatrixExponential — Type

MatrixExponential constructor method

MatrixExponential(
    a::Array{Float64,2},
    S::Array{Float64,2},
    s::Array{Float64,1},
    D::Array{Float64}=[0],
)

Inputs:

a a 1 by p Array of reals

S a p by p Array of reals

s a p by 1 Array of reals

D an optional argument, if empty then me.D is the identity, else is a p by p matrix

Throws an error if the dimensions are inconsistent.

DiscretisedFluidQueues.Mesh — Type

Mesh

Abstract type representing a discretisation mesh for a numerical scheme.

DiscretisedFluidQueues.Phase — Type

Phase

Represents a phase of the fluid queue.

Arguments:

c::Float64: the rate of change of the fluid X(t) associated with the phase

m::Int64: either -1 or 1, membership of the phase with either the set of positive or negative phases. Default is positive, i.e. 1. Only relevant for FRAPApproximation.

lpm::Bool: whether the phase has a point mass at the left/lower boundary

rpm::Bool: whether the phase has a point mass at the right/upper boundary

DiscretisedFluidQueues.PhaseSet — Type

const PhaseSet = Array{Phase,1}

A container for phases.

DiscretisedFluidQueues.PhaseSet — Method

PhaseSet(c::Array{Float64, 1}[, m::Array{Int, 1}, lpm::BitArray, rpm::BitArray])

Construct an Array of phases.

Arguments:

c: Array of rates

m: (optional) array of memberships, see Phase()

lpm: (optional) array of left point mass indicators, see Phase()

rpm: (optional) array of right point mass indicators, see Phase()

DiscretisedFluidQueues.SFMDistribution — Type

SFMDistribution{T <: Mesh} <: AbstractArray{Float64, 2}

Representation of the distribution of a DiscretisedFluidQueue.

The field coeffs is a row-vector and indexing a distribution indexes this field. Similarly, arithmetic operations on a distribution act on this vector.

Arguments:

coeffs::Array{Float64, 2}: A row vector of coefficients which encode the distribution of a DiscretisedFluidQueue and can be used to reconstruct a solution.

dq::DiscretisedFluidQueue{T}:

DiscretisedFluidQueues.SFMDistribution — Type

SFMDistribution(pdf::Function, dq::DiscretisedFluidQueue{FVMesh}, fun_evals::Int = 6)

Construct an approximation to pdf as the average of pdf of each cell.

Uses quadrature with fun_evals function evaluations to approximate cell averages.

DiscretisedFluidQueues.SFMDistribution — Method

SFMDistribution(pdf::Function, dq::DiscretisedFluidQueue{DGMesh})

Approximates pdf via polynomials.

DiscretisedFluidQueues.SFMDistribution — Method

SFMDistribution(pdf::Function, dq::DiscretisedFluidQueue{FRAPMesh}; fun_evals = 100)

Return the appropriate initial condition to approximate the initial distribution pdf for the numerical discretisation scheme defined by the dq DiscretisedFluidQueue.

i.e. for each cell compute the expected orbit position $∫pdf(x,i) ⋅ a(x) dx$ where the integral is over each cell and $a(x) = a exp(S(x-yₖ))$ if the membership of i is -1 and $a(x) = a exp(S(yₖ₊₁-x))$ if the membership of i is 1 where yₖ and yₖ₊₁ are the left and right cell edges, respectively, and a and S are defined by dq.mesh.me are parameters of a MatrixExponential.

DiscretisedFluidQueues.SFMDistribution — Method

SFMDistribution(pdf::Function, dq::DiscretisedFluidQueue{<:Mesh}, fun_evals::Int = 6)

Return a discretised version of the pdf. The method of discretisation depends on the discretisation method of dq.

Arguments:

pdf::Function: a function f(x::Float64,i::Int) where f(x,i)dx=P(X(0)∈dx,φ(0)=i) is the initial distribution of a fluid queue.

dqDiscretisedFluidQueue{<:Mesh}:

fun_evals: the number of function evaluations used to approximate f(x,i)

DiscretisedFluidQueues.SFMDistribution — Method

SFMDistribution(dq::DiscretisedFluidQueue{T})

A blank initialiser for a fluid queue distribution a distribution with coeffs=zeros.

DiscretisedFluidQueues.Simulation — Type

Simulation

A container to hold simulations.

All arguments are arrays of the same ength + a model from which the simulations came

Arguments:

t::Array{Float64, 1}:

φ::Array{Int, 1}:

X::Array{Float64, 1}:

n::Array{Int, 1}:

model::Model: