Distributions

Construction

DiscretisedFluidQueues.SFMDistribution — Type

SFMDistribution{T <: Mesh}

Representation of the distribution of a DiscretisedFluidQueue.

The field coeffs is a row-vector and indexing a distribution indexes this field. Similarly, arithmetic operations on a distribution act on this vector.

Arguments:

coeffs::Array{Float64, 2}: A row vector of coefficients which encode the distribution of a DiscretisedFluidQueue and can be used to reconstruct a solution.

dq::DiscretisedFluidQueue{T}:

DiscretisedFluidQueues.interior_point_mass — Function

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{<:Mesh})

Construct an approximation to a distribution of fluid queue with a point mass at (x,i). Returns a SFMDistribution. The method of approximation is defined by the type of <:Mesh.

Arguments:

x::Float64: The position of the point mass

i::Int: The phase of the point mass

dq::DiscretisedFluidQueue:

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{DGMesh{T}})

Constructs a polynomial approximation to the point mass at (x,i)

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{FVMesh})

Construct an approximation to a point mass at (x,i). Basically, just puts mass 1 in the cell containing (x,i).

interior_point_mass(x::Float64, i::Int, dq::DiscretisedFluidQueue{FRAPMesh})

Return the appropriate initial condition to approximate the initial distribution which is a point mass at (x,i) for the numerical discretisation scheme defined by the dq DiscretisedFluidQueue.

i.e. Compute the orbit position $a(x) = a exp(S(x-yₖ))$ if the membership of i is -1 and $a(x) = a exp(S(yₖ₊₁-x))$ if the membership of i is 1 where yₖ and yₖ₊₁ are the left and right cell edges and x∈[yₖ,yₖ₊₁], respectively, and a and S are defined by dq.mesh.me are parameters of a MatrixExponential.

DiscretisedFluidQueues.left_point_mass — Function

left_point_mass(i::Int, dq::DiscretisedFluidQueue)

Construct a SFMDistribution with a point mass at the left boundary of dq in phase i and 0 elsewhere.

DiscretisedFluidQueues.right_point_mass — Function

right_point_mass(i::Int, dq::DiscretisedFluidQueue)

Construct a SFMDistribution with a point mass at the left boundary of dq in phase i and 0 elsewhere.

Reconstructing solutions

DiscretisedFluidQueues.pdf — Method

pdf(d::SFMDistribution)

Return a function of two variables (x::Float64,i::Int) which is the probability distribution function defined by d. Approximates the density function of a fluid queue.

DiscretisedFluidQueues.pdf — Method

pdf(d::SFMDistribution, x, i)

Evaluate pdf(d::SFMDistribution) at (x,i).

DiscretisedFluidQueues.cdf — Method

cdf(d::SFMDistribution)

Return a function of two variables (x::Real,i::Int) which is the cumulative distribution distribution function defined by d. Approximates the distribution function of a fluid queue.

DiscretisedFluidQueues.cdf — Method

cdf(d::SFMDistribution, x, i)

Evaluate cdf(d::SFMDistribution) at (x,i).

Integrate time

Missing docstring.

Missing docstring for TimeIntegrationScheme. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Euler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RungeKutta4. Check Documenter's build log for details.

DiscretisedFluidQueues.integrate_time — Function

Given x0 apprximate x0 exp(Dy).

integrate_time(x0::Array{Float64,2}, D::AbstractArray{Float64,2},
    y::Float64, scheme::ExplicitRungeKuttaScheme)

Arguments

x0: An initial row vector

D: A square matrix

y: time to integrate up to

h: ExplicitRungeKuttaScheme.